+++ src/test/java/net/sf/jabref/logic/cleanup/CleanupWorkerTest.java
@@ -261,8 +261,7 @@ public class CleanupWorkerTest {
     @Test
     public void cleanupHtmlToLatexConvertsEpsilonToLatex() {
         CleanupPreset preset = new CleanupPreset(new FieldFormatterCleanups(true,
-                Collections.singletonList(new FieldFormatterCleanup("title", new HtmlToLatexFormatter(
-                        Globals.prefs.getBoolean(JabRefPreferences.USE_CONVERT_TO_EQUATION))))));
+                Collections.singletonList(new FieldFormatterCleanup("title", new HtmlToLatexFormatter()))));
         BibEntry entry = new BibEntry();
         entry.setField("title", "&Epsilon;");